trimVideo: video-payer-buttons.js -> download-trimed-video.js
