+++ client/scripts/index.js
@@ -2,6 +2,7 @@ import * as basic from "../scripts/basics.js";
 import * as notify from "../scripts/notify.js";
 import * as navigationBar from "../scripts/navigation-bar.js";
 import * as videoButton from "../scripts/video-payer-buttons.js";
+import * as trimVideo from "../scripts/download-trimed-video.js";
 import * as showAvailableVideos from "../scripts/show-available-videos.js";
 import * as currentVideoDownloads from "../scripts/current-video-downloads.js"; 
 
@@ -401,7 +402,7 @@ async function showVideo(videoSrc, videoType, videoLinkFromUrl) {
     const downloadVideoMenuContent = basic.createSection(downloadVideoMenu, "div", "vjs-menu-content");
 
     videoButton.downloadVideoButton(downloadVideoMenuContent, videoSrc, videoType);
-    videoButton.createTrimVideo(player, downloadVideoContainer, downloadVideoMenu,downloadVideoButton, downloadVideoMenuContent, videoSrc, videoType);
+    trimVideo.createTrimVideo(player, downloadVideoContainer, downloadVideoMenu,downloadVideoButton, downloadVideoMenuContent, videoSrc, videoType);
 
     downloadVideoContainer.onmouseover = function(){
       document.getElementById("downloadVideoButton").focus();
